♻️ refactor: 이미지 업로드 hook 분리
